Fix fabsf

+++ device/vertex_fit/vertex_fitter/include/VertexFitDeviceFunctions.cuh
@@ -99,7 +99,7 @@ namespace VertexFit {
     float secondBB = txB * txB + tyB * tyB + 1.0f;
     float secondAB = -txA * txB - tyA * tyB - 1.0f;
     float det = secondAA * secondBB - secondAB * secondAB;
-    if (fabs(det) > 0) {
+    if (fabsf(det) > 0) {
       float secondinvAA = secondBB / det;
       float secondinvBB = secondAA / det;
       float secondinvAB = -secondAB / det;
